Raise ArgumentError on bad args
`Terrapin::CommandLine.new` had a defualt 2nd and 3rd argument of "" and {},
respectively. But if you explicitly pass in `nil`, then you'll get an error
because `#dup` isn't on `nil`. It's not immediately appareny why you get this
error, so this change makes it so that we tell you.

Fixes https://github.com/thoughtbot/cocaine/issues/98