Loosen dependency lock on ClimateControl
ClimateControl's 1.0 release doesn't have any breaking API changes
(the major version bump is a reflection of dropping support for some
older Rubies). This change loosens the dependency lock for
`climate_control` to allow any version in the 1.x series to be used.
This allows Terrapin and kt-Paperclip to coexist with the new
ClimateControl release.

ClimateControl 1.0 release notes:
https://github.com/thoughtbot/climate_control/releases/tag/v1.0.0