Fix the stderr test
The `BackticksRunner` and `PopenRunner` do not support outputs on `stderr`,
therefore the spec that verifies whether large output on `stderr` hangs or
not, doesn't make sense for these runners.

The only reason it works on MRI is because the way the spec is written,
selects the `ProcessRunner`.

On JRuby the `ProcessRunner` is not supported, so it falls back to the
`BackticksRunner` and then makes the spec fail.